Add course id to mentor
- Add new course id to mentor who create it
- Remove conflict from app.js
- The conflict was duplication of using upload package

+++ middlewares/coursesErrorHandling.js
@@ -46,14 +46,19 @@ exports.coursesErrorHandling = (err, req, res, next) => {
         req.filesPath.coursePicture = '';
       }
   
-      // delete course media
-      if(req.filesPath.mediaUrls.length != 0) {
-        for(let i = 0; i < req.filesPath.mediaUrls.length; i++) {
-          fs.unlinkSync(`${__dirname}/../public${req.filesPath.mediaUrls[i]}`);
+      // check if there's  mediaUrls
+      if(req.filesPath.mediaUrls) {
+        
+        // delete course media
+        if(req.filesPath.mediaUrls.length != 0) {
+          for(let i = 0; i < req.filesPath.mediaUrls.length; i++) {
+            fs.unlinkSync(`${__dirname}/../public${req.filesPath.mediaUrls[i]}`);
+          }
+  
+          // clear pathes after deleting files
+          req.filesPath.mediaUrls = [];
         }
 
-        // clear pathes after deleting files
-        req.filesPath.mediaUrls = [];
       }
 
     }
